requirements.txt
# HTTP通信
aiohttp>=3.8.0
requests>=2.31.0

# Googleスプレッドシート連携
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0

# 環境変数管理
python-dotenv>=1.0.0

# YAML処理
PyYAML>=6.0.0

# 画像処理
Pillow>=10.0.0

# 非同期処理
asyncio>=3.4.3

# テスト
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# セキュリティ
cryptography>=41.0.0
argon2-cffi>=21.3.0

# キャッシュ管理
redis>=4.5.0

# ログ管理
python-json-logger>=2.0.7

# 型ヒント
typing-extensions>=4.0.0

# ユーティリティ
python-dateutil>=2.8.2
pytz>=2023.3

# 追加の依存関係
psutil>=5.9.0
opencv-python>=4.8.0
numpy>=1.24.0 